## Refactiorization of sprite animations. Initial commit.
### New classes.
`SpriteAnimation` class has beed introduced. It's purpose is to provide 
some generalization to the sprite animations concept, that would allow 
some more flexibility than the default `libgdx` `Animation` class. It 
takes a `TextureRegion` and internally prepares `Animation` instance.

`AnimationWalk` extends `SpriteAnimation` and provides an update method, 
that adjusts position of the sprite to be drawn to emulate movement.

For more information please read documentation.

### `Person`:
An example of `AnimationWalk` is introduced in the `draw_walk` method.

### General changes.

Changed units of variables related to screen position from `int` to 
`float` as it's the natural unit in libgdx.

+++ Strategix/src/com/me/mygdxgame/Marker.java
@@ -281,8 +281,8 @@ private void draw_strzalki(SpriteBatch spriteBatch)
 	int correctorX = 8;
 	int correctorY = 8;
 	Vector2 nadglowa = new Vector2(24,48);
-	int isoX = position.getIsoX();
-	int isoY = position.getIsoY();
+	float isoX = position.getIsoX();
+	float isoY = position.getIsoY();
 	
 	if(strzalka == Direction.DOWN){spriteBatch.draw(strzalki[1], isoX+correctorX+nadglowa.x, isoY-correctorY-1+nadglowa.y);}
 	else{spriteBatch.draw(strzalki[1+4], isoX+correctorX+nadglowa.x, isoY-correctorY+nadglowa.y);}
